msm: kgsl: Do not use uint64_t in header included by user space
Do not use uint64_t in header file that can be included by user
space application as the user space application may not have
uint64_t defined.

+++ include/uapi/linux/msm_kgsl.h
@@ -769,7 +769,7 @@ struct kgsl_perfcounter_query {
 struct kgsl_perfcounter_read_group {
 	unsigned int groupid;
 	unsigned int countable;
-	uint64_t value;
+	unsigned long long value;
 };
 
 struct kgsl_perfcounter_read {
